[Go] Improve cache handling for long-running requests. (#31)
This change reworks the duplicate-request cache to properly handle
in-flight requests (e.g. requests that have successfully started but not
yet completed processing). Any duplicate requests that come in while the
original request is in-flight will be dropped, since the completion of
the original request will also resolve all of them.

In addition, the cache now tracks duplicate requests even for requests
that hit the equivalent-request cache, to ensure that duplicate requests
will always return the same result. All requests that refer to the same
response share the same payload in memory, which is reference-counted to
track when the memory is actually released (for purposes of cache
trimming).