[Go] Align state-store API with feedback. (#75)
This change aligns the Go state-store API with decisions brought up in
the Rust API review:
* DEL and VDEL now return the number of items deleted (for
future-proofing and consistency with the service).
* The `VDel` method is now cased appropriately to its meaning.
* Well-known service errors are now provided as enum values.
* The client now takes generic type parameters for its key and value
types. This has two benefits for Go:
- It can avoid inconvenient casting when providing the keys and
(primarily) values.
- While `string` and `[]byte` are mostly isomorphic, slight behavioral
differences cause Go to perform a copy when values are cast between
types. By retaining the original type as long as possible, we avoid
bouncing between types and the corresponding incurred cost.