[Rust] ManagedClient (#76)
# `azure_iot_operations_mqtt` Changes
* The functionality of `MqttProvider` trait has been replaced by the
`ManagedClient` trait.
* The `MqttPubReceiver` trait has been renamed to `PubReceiver` for
consistency with the new `ManagedClient` trait and to establish a clear
difference between MQTT primitive traits and higher-level client traits
* The `ManagedClient` trait now defines a struct that implements
`MqttPubSub` and provides the ability to create a `PubReceiver`
* The `Session` does not implement `ManagedClient` like it previously
did `MqttProvider`. Instead, it now has the ability to spawn new
instances of a `SessionManagedClient` which implements `ManagedClient`
and thus, by extension. also `MqttPubSub`
* The `SessionPubSub` has been removed, as its functionality is covered
by the `SessionManagedClient`
* Samples have been updated to reflect these changes
* `Session::get_session_exit_handle()` has been renamed to
`Session::create_exit_handle()` for consistency with the new
`Session::create_managed_client()`
* Various other internal refactoring

# `azure_iot_operations_protocol` Changes
* Implementations of envoys have been updated to use the `ManagedClient`
trait
* Factory methods now consume a struct implementing `ManagedClient`
instead of borrowing a struct implementing `MqttProvider`
* Samples have been updated to reflect these changes

# Implications
* Providing MQTT client functionality in an application is much more
flexible - you can simply give a `SessionManagedClient` instance to any
component that requires MQTT at any time instead of having to pass
around the `Session` itself.
* As `SessionManagedClient` is cloneable, this means you can store one
to use/clone later while the `Session` is running (before, you could not
gain additional `SessionPubSub`/`SessionPubReceiver`s). As a result,
when using the `protocols` library, you can dynamically add
executors/invokers during your application.
* Additionally, the static dispatching of all envoys has been simplified
to only dispatch over their unique data types + the struct implementing
`ManagedClient` which is much more readable.
* As the new `SessionManagedClient` contains state from the `Session`
internally, this lays the groundwork to build a notification system for
exit/shutdown scenarios in the envoy implementations (although the
necessary APIs have not yet been added to the `SessionManagedClient`)
* Similarly, these changes lay the groundwork for fixing the
bottenecking issues in the receive flow (as the infrastructure exists
there to enable us to delay the creation of communication channels until
we have negotiated a connection with the broker and know the max PKID)
* Both `azure_iot_operations_protcol` and `azure_iot_operations_mqtt`
will need to be version bumped on the next release now as this PR
contains breaking changes for both.

# Notes/Caveats
* This PR is focused on making the API surface the way we want it.
Internally, there still ought to be additional refactoring in response
to these changes, particularly around the `PubDispatcher` which is now
being shared in a way it was not designed to be.
* Related to the above, there's the (small) potential of a race
condition that has been created with the `PubDispatcher`. I wanted to
fix it here, but it would have ballooned the scope, so that will be a
separate PR. It is unlikely, and relatively harmless (small chance that
under certain conditions a redelivered message from a previous MQTT
session (i.e. clean_start=false) may be not delivered to the
application).

---------

Co-authored-by: Valerie Avva Lim <54871851+vaavva@users.noreply.github.com>