enable case sensativity in markdown, and dont check code blocks (#85)
Unfortunately, Cspell doesn't allow different dictionaries to be used in
different parts of a file. The dictionaries are compiled together for
each file. This means for markdown there are restrictions in that you
can either prioritise the text by enabling case sensitivity and using an
english dictionary, or prioritise the code blocks by disabling case
sensativity and added additional programming language dictionaries.

This PR prioritises the text.

Signed-off-by: Ryan Winter <ryanwinter@outlook.com>