[Rust] State Store Get/Set/Del/VDel & Sample (#74)
This PR:
- Adds infrastructure for Azure IoT Operations Services, for now just
the State Store
- Adds an implementation of basic State Store operations (Get, Set, Del,
and VDel) that builds on top of the RPC Command Invoker protocol.
- Adds Serialization and deserialization for RESP3 State Store payloads
- Adds a simple sample to show how to use the new State Store APIs
- Adds unit tests for RESP3 serialization/deserialization
- Adds unit tests for the State Store Client
- Updates Hybrid Logical Clock implementation to be more generic to work
with the State Store service
- aligns package versions across the workspace (mqtt, protocol,
services) to keep from having to specify package versions repeatedly

This PR does not:
- add implementations for state store APIs around key notification
- add network connected tests for Get/Set/Del/VDel and all of their
permutations (I am working on this though!)

See #63 for more details about the API surface. As that doc finalizes,
there may be some surface level changes to this PR, but the
implementation logic should stay fairly stable.