[Go] Replace Listen/done pattern with split Start/Close pattern (#121)
With the listener handler instantiation moved to the respective
constructors (#108), there are now goroutines that need to be cleaned up
even if `Listen` is never called. This change replaces the "done"
callback returned by `Listen` with an explicit `Close` that frees all
active resources from the listener (and can be called independent of
`Listen`).