[Rust][MQTT][Services] .kind() method for top-level error structures (#184)
It was impossible to match on `SessionError` and `StateStoreError` due
to being opaque. Exposed a method to allow matching on the inner enum.

Arguably this is not an ideal solution either - there's still a lot of
strange things about error in my opinion, and I'm not convinced structs
with enum kinds provide meaningful value. I'm going to schedule a
meeting with one of the Azure Rust experts on this topic specifically to
follow-up on earlier conversations about error design.

In the meantime, this will unblock some customers.

NOTE: This PR necessitates a version bump for both `mqtt` and `services`