[Go] Move generated README.md to API.md (#244)
The default behavior of `gomarkdoc` (generating the API documentation
directly in the `README.md`) results in the API documentation being
repeated on pkg.go.dev (as it appears in both the "README" and
"Documentation" sections). This change therefore moves the generated API
docs to `API.md` in preparation for manually-written `README.md` files
(which will simply link to the API docs). This will also make the Go
documentation more akin to other languages in this repo.