Add DataSchema prop to TelemetrySender (#236)
This PR adds a `DataSchema` property to the TelemetrySender and uses its
value for the value of the `dataschema` property in a CloundEvent.

The ProtocolCompiler sets the value of the new `DataSchema` property as
follows:

* For an Interface with combined telemetries (no "{telemetryName}" token
in the topic pattern), the data schema is the Interface DTMI. E.g.,
`dtmi:test:mySensor;1`
* For an Interface with separated telemetries, the data schema of each
individual Telemetry is the DTMI of the Telemetry element within the
Interface. E.g., `dtmi:test:mySensor:_contents:__Temperature;1`

The METL language is enhanced to support setting the `DataSchema`
property, and two METL test cases are updated to test its use in
CloudEvents.