[Rust] Codegen samples and template updates for linting (#246)
* Added a `sample_applications` directory to the Rust SDK, containing a
codegen version of the `counter_client` and `counter_server` previously
that were samples in our `protocol` crate.
* Updated codegen templates to produce code that passes linting

**Notes**
Wrt the sample applications, I do not think this is the ideal layout -
arguably, the `counter_client` and `counter_server` should be `examples`
of the generated crate, which would involve less boilerplate
infrastructure due to examples being implicitly crates (rather than
explicit ones as we have here), and would allow these samples to be used
like the `examples` in any other crate (via the use of the `--example`
parameter to cargo), but depending on the eventual strategy for
long-term codegen maintenance, may be problematic.

There are also other strange artifacts of inconsistent approach, like
how even though the DTDL code is maintained locally, it's generated
dymanically for the actual pipeline tests. Again, this needs a unified
approach across language.

For now, I'm just trying to stick to what is currently being done in the
other languages as much as possible.