Fix Cross-Language E2Es (#321)
This pipeline wasn't validating everything expected. The problems:
1. If codegen failed to generate, this didn't fail the pipeline, so the
tests would be run against whatever was currently checked in
2. The parameters provided to codegen were intended for debug mode (to
be able to test local copies of the SDKs), and the parameter for the
modelFile location was incorrect in Go
3. Once Rust was actually getting generated, an issue popped up where
cargo dependencies were being installed in the background, and not
finishing before the next Rust step, which then caused intermittent
failures of Rust tooling
4. CI didn't validate whether checked in generated code was staying up
to date
5. (different pipeline) Rust CI wasn't getting kicked off for all
changes that might affect it

This PR fixes:
1. Thanks JD and Tim for helping me figure the changes for this one out
in these PRs (#317 and #316). I've folded the fix into this PR.
2. I switched to using codegen in Debug mode, so that we are validating
with the latest codegen and the latest SDK code.
3. Added a Rust setup step to pre-initialize Rust tools. As part of
this, split out running codegen for the different languages so that they
all matrix runs don't have to run the rust initialization step if it
isn't relevant
4. The pipeline checks if generated code is different than what is
checked in and fails the run if it isn't (**NOTE:** I'm not sure if this
will cause problems if the folks who run on windows check in new
codegenerated files. If that's the case, there might be some git
settings they can change to have line endings conform properly, or this
check might need to be removed.)
5. Updated triggers for Rust CI to run on codegen changes and
statestore-cli changes

---------

Co-authored-by: John (JD) Douceur <johndo@microsoft.com>