[Rust][MQTT] AckToken API for acknowledgement (#338)
- Modified the signature of `PubReceiver`'s `.ack()` method to use a new
`AckToken` structure for acking, and removed the old approach using the
`MqttAck` trait
- Updated the `protocol` crate to use the new style
- Added a new sample demonstrating the use of the `AckToken` API for
manual acking.

NOTE that this is very much a transition PR.
- The new API is added, but the implementation is not finished
underneath. It only has minimal viable implementation to retain current
functionality.
- The consuming code in `protocols` is the minimal viable implementation
to retain functionality as well. With the new flexibility the API
provides, this can and should be restructured to be simpler.
- The signature of the `PubReceiver` `recv` method will be renamed to
`recv_manual_ack` in a follow-up PR, with a different `recv` being
introduced. This has been done to make the transition easier.