[E2E] have git diff fail if there are differences (#370)
the git diff check to verify that what is codegen'd in the pipeline is
the same as what is checked in wasn't failing properly. This PR fixes
the check to fail properly and commits the changes in the repo that are
missing. This PR also fixes the Rust validations to work with the new
codegen header