Fix correlation ID check in Go METL executor awaitPublish (#378)
By code inspection, I noticed that the implementation of awaitPublish in
the CommandExecutorTester was incorrect, and I started this PR to fix
it. I believe the test code is now correct, but a consequence is that
the METL test cases for caching equivalent requests are now failing.

I think what is happening is that the cache is storing the entire
response message, including the correlation ID, and this is getting
returned from the executor even for equivalent requests (which have
different correlation IDs). I think the code should perhaps be
refactored so that the cache stores only the response, and the
CommandExecutor calls `build` on the response retrieved from the cache,
setting the correlation ID (and message expiry and response topic) from
the current message, rather than getting these values from the cache.