normalize dotnet METL cs file line endings (#388)
A bunch of the .NET METL tester files have mixed line endings. I don't
know how they got this way; Git should be automatically regularizing so
that the endings are always LF in the repo and file-system-appropriate
on a local dev system. I'm hoping this will be a one-time fix, and the
endings will stay consistent from now on.