Remove ProtocolCompiler.IntegrationTests (#413)
These tests are quite old and have badly rotted. They are not currently
used, and it would likely take more effort to make them work than it
would take to rearchitect them, particularly given some of the changes
we have made to the ProtocolCompiler.

Their mere existence is getting in the way of a big refactoring of the
ProtocolCompiler I have in progress, so I'd like to just get rid of them
to help move this rework along.