[Go] ADR 10 & 11: Update cloud events (#418)
This change replaces the `CloudEvent` field on the `TelemetryMessage`
struct with a `CloudEventFromTelemetry` function. This makes it less
coupled with the API and allows us to return errors when cloud events
are explicitly requested but invalid. The `WithCloudEvent` option
remains on the sender for convenience and to allow defaults to be
specified properly at runtime.

This change also updates the cloud events to correctly use the
`ContentType` message property for the `DataContentType` field (and will
error if one is specified and it does not match the serializer).