[Rust][MQTT] MockClient call sequencing (#422)
- Added the ability to track publishes, subscribes, unsubscribes and
acks with the `MockClient` in `interface_mocks.rs`

**Notes**
- It does seem the direction this module is trending is not towards a
truly mocked "interface" (e.g. a pure `PubSub` mock), but rather bespoke
monolithic mocks (`MockClient`) but I won't rename it yet.
- There is some overlap here with some of the METL testing
infrastructure included in the `protocol` crate. It probably would be
best to consolidate this at some point, but that is out of scope for now
as this PR largely serves to unblock #416