Implement ADR 13, Language-Appropriate Generated Names (#425)
This change started out innocently enough as an implementation of ADR
13, but it snowballed into a multi-day effort of refactoring the
ProtocolCompiler. The naming logic was scattered all over the codebase,
so cleanly implementing new naming rules required centralizing the
naming logic. The naming rules specified in
[ADR13](https://github.com/Azure/iot-operations-sdks/blob/main/doc/dev/adr/0013-codegen-generated-names.md)
are now codified in a new
[CodeName](https://github.com/Azure/iot-operations-sdks/blob/johndo/plumb-CodeName/codegen/src/Azure.Iot.Operations.ProtocolCompiler/CodeName.cs)
class. To cleanly support Raw serialization, I added class
[RawTypeName](https://github.com/Azure/iot-operations-sdks/blob/johndo/plumb-CodeName/codegen/src/Azure.Iot.Operations.ProtocolCompiler/RawTypeName.cs),
and to support empty command requests and responses, I added
[EmptyTypeName](https://github.com/Azure/iot-operations-sdks/blob/johndo/plumb-CodeName/codegen/src/Azure.Iot.Operations.ProtocolCompiler/EmptyTypeName.cs).
All three classes export the new
[ITypeName](https://github.com/Azure/iot-operations-sdks/blob/johndo/plumb-CodeName/codegen/src/Azure.Iot.Operations.ProtocolCompiler/ITypeName.cs)
interface.

With the exception of the three new classes mentioned above, nearly all
of the changes herein are either (a) plumbing, (b) calls to methods of
the new classes, (c) changes to generated code, or (d) changes to code
that interacts with generated code and therefore uses the new names.