[Rust][Services] State store shutdown (#426)
This PR updates the shutdown method/flow for the State Store Client to
account for changes in
https://github.com/Azure/iot-operations-sdks/pull/361 and
https://github.com/Azure/iot-operations-sdks/pull/338

Context:
In https://github.com/Azure/iot-operations-sdks/pull/361, the shutdown
flows for the envoys were completed. The state store client's shutdown
flow was not updated to reflect similar changes.

Changes:
- recv_cancellation_token is changed to a Notify (shutdown_notifier).
This allows the recv loop to shutdown the telemetry receiver when the
state store client is shutdown, but allows telemetry messages to
continue to be received until there are no more.
- side effect: since telemetry_receiver.shutdown() is a failable
operation (note that only the unsubscribe can fail, the mqtt receiver
will still always be closed so recv will still eventually return None,
so the receive loop will still always end), we would want to retry the
shutdown so that the unsubscribe can succeed. However, it's possible for
this to continue to try forever, so I added a max of 3 attempts for this
operation. The consequence of this is that it is possible for the
unsubscribe to never succeed, but the task will always be able to end.
- notifies the shutdown_notifier on drop to clean up the
telemetry_receiver. Note that nothing needs to be done for the command
invoker as it's drop will be executed when it is dropped as part of the
state store client's drop.
- tokio-util dependency removed from Services crate since it is no
longer used
- network test added to confirm that `None` is returned to the
notification receiver after shutdown is called.