Revise METL to remove assumption of JSON serialization (#461)
## Context for the change

* Serializers are not part of the SDKs and should not be assumed by the
METL language or test infrastructure.
* New test cases are needed (but not yet written) that will require more
control in METL than currently available.

## Substance of the change

This PR modifies the METL language as follows:

* It adds a 'serializer' property to elements of the 'executors',
'invokers', 'receivers', and 'senders' arrays. Each serializer is
controlled by a set of [configuration
properties](https://github.com/Azure/iot-operations-sdks/blob/johndo/metl-no-serialization/doc/dev/generated/MetlSpec.md#executorserializer).
* It removes the 'bypass-serialization' property from actions 'receive
request', 'receive response', 'receive telemetry'.

This PR modifies the METL interpreter testers in C#, Go, and Rust to
implement the METL language changes described above.

This PR also makes minor changes to a few test cases to use the new
'serializer' feature instead of the old 'bypass-serialization' feature.
For example, test case
[CommandExecutorRequestMalformedPayload_RespondsError](https://github.com/Azure/iot-operations-sdks/blob/main/eng/test/test-cases/Protocol/CommandExecutor/CommandExecutorRequestMalformedPayload_RespondsError.yaml)
is modified so that instead of passing in malformed JSON text with
'bypass-serialization' set to true, it sets the 'fail-deserialization'
on the executor's serializer to true.

### Implementation note for Rust

The Rust SDK does not configure each executor, invoker, receiver, and
sender with a serializer instance. Instead, the serialization and
deserialization code is defined on the type of struct that is being
serialized or deserialized. For serialization, this does not present a
significant problem because `serialize` is implemented as a method on
the struct. However, `deserialize` is a static method on the type, which
incurs the usual test problems of static methods, in particular making
it impossible to set specific behavior using runtime values read from
test cases.

To work around this, I introduced the following horrible hack. The
`TestPayload` class implements the `PayloadSerialize` trait, and for
testing serialization, the test infrastructure sets values in the
`TestPayload` instance that control the behavior of the `serialize`
method. For testing deserialization (hold your nose), the test
infrastructure sets values in the `TestPayload` instance, then
serializes the `TestPayload` into bytes using JSON serialization, then
uses its `deserialize` method to first JSON-deserialize the bytes it is
given, after which it looks at the deserialized configuration values to
determine how to behave. Erp, sorry.

The above approach clearly fails if there is more than one serializer
configuration per test case. So, I added a new test platform feature,
'multiple-serializers', which is required for any test case that needs
multiple components that have different serializer configurations, and
the Rust tester indicates that it does not support this feature. In
practice, I suspect there will never be any test cases that use this
feature, so this should be a non-issue modulo its general hackiness.