[Rust][MQTT][Protocol] Shared subscription support (#465)
# Context

Using shared subscriptions via the `service_group_id` has been an option
for the command executor but the underlying infrastructure to support it
was missing.

# Changes
This PR adds support for shared subscriptions in the MQTT and Protocol
crates

Protocol:
- Modifies `TopicPattern` to take an optional share name. If it is
present, it will return the full shared subscription topic filter when
obtaining the topic to subscribe.
- Invoker: Now adds `$partition` as a user property on every invoke. It
contains the client id.
- Adds example to demonstrate usage of shared subscriptions. Can be used
against the simple invoker sample.

Session:
- Modified `topic.rs` :
- Adds validation for shared subscription topic filters for topic names
(not allowed) and topic filters.
- Modifies topic matching function to account for shared subscription
topic filters.

# Testing
- Tested shared subscription manually
- Added unit tests to `topic_processor.rs` and `topic.rs`