Enable ProtocolCompiler to generate code from schema files w/o DTDL (#495)
Context: @koepalex has a use case for generating code from JSON Schema
files that are not derived from a DTDL model.

This change allows the ProtocolCompiler to be invoked without a
--modelFile or --dmrRoot option, in which case it scans the working
directory for schema files from which to generate code. The gen.sh files
under codegen/test/samples illustrate the behavior.