Make Rust CountdownEventMap async (#507)
The CountdownEventMap in the METL tests folder was implemented with a
condition variable, which blocks the thread while waiting. This caused a
test hang in the single-threaded async framework used by the METL tests.
This change replaces the condition variably with a Notify, which is
designed for async usage.

With this change, test case
CommandExecutorRequestsCompleteOutOfOrder_RequestAckedInOrder now
passes, so it is removed from the list of PROBLEMATIC_TEST_CASES.