Force legalize project name derived from folder name (#504)
This changes the PoR for how the ProtocolCompiler derives a project name
from the output folder name. In most cases, this change will have no
effect. But for folders that contain characters that are generally not
legal as names in programs, this change will derive a legal name that
contains only alphanumerics and underscores, with the first character
non-numeric.

This change is to address a problem that arose this morning caused by an
output folder name in kebab-case, which resulted in a project name that
was illegal in an AVSC file. The error behavior was not easy to
understand.