Normalize line endings (#518)
This repo has been seeing a lot of churn caused by inconsistent line
endings. Among other problems, this makes it difficult to review changes
in PRs that appear to completely rewrite files when in fact only a small
portion of the file has changed.

This PR:

* modifies the root .gitattributes file to set auto line endings for all
text files other than shell scripts (*.sh), which will continue to be LF
always for Docker compatibility
* renormalizes all files per the newly set .gitattributes