[Rust][Protocol] TopicPatternError initial version (#519)
Initial version of `TopicPatternError` and `TopicPatternErrorKind` to
allow for more generic handling of errors that stem from topic patterns.

Known limitations:
- The current translation to the existing `AIOProtocolError` involves
some kind of janky functions that will be replaced when
`AIOProtocolError` is replaced with more tightly scoped errors
- The error design for `TopicPatternError` is not ideal, and
`topic_processor.rs` likely needs to be refactored to disentangle
concepts of share names, namespaces, token replacements and topic
patterns themselves, but this is the best error I could implement 1:1
for the existing design and scope.

---------

Co-authored-by: Raul Leclair <raulleclair@microsoft.com>