[.Net] Apply Exponential Backoff Retry Policy to the leader election process  (#479)
This pull request introduces several changes to the retry policies and
configurations in the Azure IoT Operations Protocol. The most
significant changes involve enhancements to the
`ExponentialBackoffRetryPolicy` class, the addition of a new
`ExponentialBackoffRetryPolicyWithExpiration` class, and updates to the
`LeaderElectionClient` and `LeasedLockClient` classes to incorporate
these new retry policies.

### Enhancements to Retry Policies:

*
[`dotnet/src/Azure.Iot.Operations.Protocol/Retry/ExponentialBackoffRetryPolicy.cs`](diffhunk://#diff-f97efe7b1b8369b0c8cf0c130e6a5f8e02f788713b58e0a4d7c6203324234914L9-R68):
Updated the `ExponentialBackoffRetryPolicy` class to provide more
detailed documentation, include a base exponent for backoff
calculations, and add a new constructor to support different
initialization parameters.
*
[`dotnet/src/Azure.Iot.Operations.Protocol/Retry/ExponentialBackoffRetryPolicyWithExpiration.cs`](diffhunk://#diff-cd9a9e23429f22bac73113f528e307fca768712b71792f988b2bea57d9321923R1-R77):
Added a new class `ExponentialBackoffRetryPolicyWithExpiration` that
extends the `ExponentialBackoffRetryPolicy` to include an expiration
interval after which retries will no longer be attempted.

### Updates to Client Classes:

*
[`dotnet/src/Azure.Iot.Operations.Services/LeaderElection/LeaderElectionClient.cs`](diffhunk://#diff-b46aa354fb4a526ef315fb12f1df5c012b5f8b4c07dba330312faf8131892102R132-R149):
Modified the `LeaderElectionClient` class to accept a retry policy
parameter, with a default policy using the new
`ExponentialBackoffRetryPolicyWithExpiration`.
*
[`dotnet/src/Azure.Iot.Operations.Services/LeasedLock/LeasedLockClient.cs`](diffhunk://#diff-a0b2f6fa562e5bfc738f5b306f20770db57441e2c6d22cf2b59f28ca0a4b7cfbR110-R114):
Updated the `LeasedLockClient` class to accept a retry policy parameter
and implemented the retry logic to prioritize previous leader renewal.
[[1]](diffhunk://#diff-a0b2f6fa562e5bfc738f5b306f20770db57441e2c6d22cf2b59f28ca0a4b7cfbR110-R114)
[[2]](diffhunk://#diff-a0b2f6fa562e5bfc738f5b306f20770db57441e2c6d22cf2b59f28ca0a4b7cfbR338-R349)

### Minor Fixes and Improvements:

*
[`dotnet/.editorconfig`](diffhunk://#diff-cfb867d5de3aaa02df169c9665b9d05d4f96b045d92063289b1e1e9b2d88b2cfL97-R96):
Updated the naming style rule for `first_word_upper_case_style` to use
`pascal_case` instead of `first_word_upper`.
*
[`dotnet/src/Azure.Iot.Operations.Protocol/Retry/IRetryPolicy.cs`](diffhunk://#diff-67f012295c600254de059073df09536db6bd27bdcab31e0928464da4f692f315L29-R29):
Updated the `IRetryPolicy` interface to allow a nullable `Exception`
parameter in the `ShouldRetry` method.
*
[`dotnet/src/Azure.Iot.Operations.Protocol/Retry/NoRetryPolicy.cs`](diffhunk://#diff-17927ab802afa78637340f4f2da88c81cdb45c2f3b224d2e4f1cd781480723f4L21-R21):
Updated the `NoRetryPolicy` class to allow a nullable `Exception`
parameter in the `ShouldRetry` method.