[Rust][MQTT] Mitigating session exit handle returning broker unavailable (#521)
# Context

When calling `try_exit` from a session exit handle while connected, it
would sometimes return Ok(()) and other times return an error. This
inconsistent behavior was caused by a select statement that alternated
between checking if the client had already exited and checking if the
client was disconnected. If the client was disconnected, it would return
an error since it would assume not being connected to the broker meant
it was unavailable rather than that a user initiated disconnect had
happened.

# Changes
Biases the aforementioned select loop within try exit to always check if
the client exited first before checking if the client is disconnected
and as such return an error as if the broker was unavailable.

Note: A complete fix for this issue would have to occur at the mqtt
client level which can sometimes falsely report disconnects as a result
of a network failure rather than a user initiated disconnect.