[Rust][Protocol] Fixing rounding error when determining response expiry time in executor (#527)
# Context

After receiving a request, the executor calculates the expiry time of
the response based on the time elapsed from when the request was
received to when the response is sent. For example, if the request is
received at time 0 with a message expiry of 10 seconds and the
application takes 5 seconds to process, the executor will send a
response with a message expiry of 5 seconds. There was an issue with
calculating this final expiry time because the seconds portion of the
Duration value tracking it was used, which would result in a rounding
down of the second.

# Changes
- The response message expiry time is now rounded up. 
- Minor changes to improve the readability of the executor.
- This change fixes the `CommandExecutorRequest_TimeoutPropagated` test,
so it is now enabled.