[.NET] Add sample and integration tests that demonstrate custom topic tokens in DTDL + generated code (#415)
This PR adds a new DTDL file that has 1 telemetry definition and 1 RPC
definition. In the topic strings for both is a single custom topic
token.

The DTDL output is added to the existing test/sample DTDL files (math,
counter, greeter, etc.) in our TestEnvoys and is added to the
SampleClient/SampleServer projects like those other DTDL outputs already
are.

---------

Co-authored-by: timtay-microsoft <timtay-microsoft@users.noreply.github.com>