[Rust][Sample Applications][Services] Fixing executor ID being set overwrites custom token map and viceversa in codegen (#547)
# Context

When creating a request in codegen, if the topic has the standard token
`{executorId}` an additional option to set this specific field will be
exposed. If this is used in conjunction with the already existing
`topic_tokens` function used to replace tokens on an invoke, the HashMap
with the executor ID value will be overwritten and viceversa.

# Changes

Adds a `topic_tokens` private field to add topic tokens to rather than
overwriting.