Fix Go Enum codegen (#569)
While working on sample code for ADR 19, I ran into a code-generation
bug relating to Go enums. The ProtocolCompiler was generating this:

```go
type SpeedSchema int32

const (
	Fast SpeedSchema = 1
	Slow  = 2
)
```

which was causing `Slow` to have a type of `int` instead of
`SpeedSchema`.

The present PR fixes this so that it generates:

```go
type SpeedSchema int32

const (
	Fast SpeedSchema = 1
	Slow SpeedSchema  = 2
)
```